+++ tnnt_discordbot_cogs/auth_hooks.py
@@ -24,4 +24,5 @@ def register_cogs():
         "tnnt_discordbot_cogs.cogs.price_check",
         "tnnt_discordbot_cogs.cogs.recruit_me",
         "tnnt_discordbot_cogs.cogs.welcome",
+        "tnnt_discordbot_cogs.cogs.where_is",
     ]